Did some gesty do's

+++ ChronoLoop/ChronoLoop/Core/TimeManager.cpp
@@ -320,7 +320,7 @@ namespace Epoch {
 				return;
 			}
 
-			if (mtempCurSnapFrame > mTimeline->GetCurrentGameTimeIndx())
+			if ((unsigned int)mtempCurSnapFrame > mTimeline->GetCurrentGameTimeIndx())
 				return;
